pyproject.toml
[project]
name = "projeto-markowitz"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Mateus Lopes Albano", email = "50590861+mateus-lopes@users.noreply.github.com"},
    {name = "mateus-lopes", email = "mateusalbano22@gmail.com"},
]
dependencies = [
    "streamlit>=1.33.0",
    "streamlit-authenticator>=0.3.2",
    "pandas>=2.2.2",
    "numpy>=1.26.4",
    "matplotlib>=3.8.4",
    "seaborn>=0.13.2",
    "pillow>=10.3.0",
    "rpds-py>=0.18.0",
    "pyarrow>=16.0.0",
    "fastapi>=0.112.0",
    "uvicorn>=0.30.5",
    "sqlalchemy>=2.0.32",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.2",
    "python-decouple>=3.8",
    "passlib>=1.7.4",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.9",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = true

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"

